+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 * Support X-Job-Realm for use
 * Support Administrator realm token
 * Tests: enhancing Test environment to work with lizmap 3.6
+* Job label based on input value
 
 ## 0.2.0 - 2022-11-22